fix(ui5-option-group): remove unrelated inherited public APIs
ui5-option-group extended ui5-li-group and therefore inherited a number
of public APIs that are irrelevant to grouping options inside a
ui5-select and were never wired into its template:

- headerAccessibleName property
- wrappingType property
- header slot
- move and move-over events
- header and title CSS parts

Extract an abstract ListItemGroupBase that holds only the shared group
contract (headerText, the default items slot, and the isListItemGroup /
groupHeaderItem hooks the internal ui5-list relies on to flatten grouped
items). Both ui5-li-group and ui5-option-group now extend this base, so
the drag-and-drop, wrapping and header-slot APIs live solely on
ui5-li-group where they belong, while ui5-option-group exposes only
headerText and its options slot.

ui5-li-group's public API is unchanged.

Also add a standalone documentation page for ui5-option-group.

Fixes #13992
Fixes #13991
