+++ packages/website/docs/_components_pages/main/Select/OptionGroup.mdx
@@ -1,0 +1,7 @@
+---
+slug: ../../OptionGroup
+---
+
+<%COMPONENT_OVERVIEW%>
+
+<%COMPONENT_METADATA%>
